feat: delete posts from my posts view
