+++ components/posts/PostFullView.tsx
@@ -1,10 +1,10 @@
-import { formatDate, formatValue } from "@/utils/posts";
+import { formatValue } from "@/utils/posts";
 import { type User, type Post } from "@prisma/client";
 import EditableTitle from "./EditableTitle";
 import EditableBody from "./EditableBody";
 import EditableDescription from "./EditableDescription";
 import PostFooter from "./PostFooter";
-import DeletePost from "./DeletePost";
+import PostPreHeader from "./PostPreHeader";
 
 interface PostFullProps {
   post: Post;
@@ -13,7 +13,6 @@ interface PostFullProps {
 }
 
 export default function PostFullView({ post, user, slug }: PostFullProps) {
-  const formattedDate = formatDate(post.updatedAt);
   const content = formatValue(post.content);
   const isEditable = !slug ? post.ownerId === user.id : false;
 
@@ -28,15 +27,7 @@ export default function PostFullView({ post, user, slug }: PostFullProps) {
         key={post.id}
         className="[&:not(:last-child)]:border-b-1 border-gray-200"
       >
-        <div className="flex items-center justify-between mb-4">
-          <time
-            dateTime={formattedDate}
-            className="text-gray-500 italic text-xs"
-          >
-            {formattedDate}
-          </time>
-          <DeletePost postId={post.id} />
-        </div>
+        <PostPreHeader post={post} />
         <div className="group relative">
           <header>
             <EditableTitle title={post.title} postId={post.id} />
