+++ components/posts/PostPreview.tsx
@@ -1,7 +1,7 @@
-import { formatDate } from "@/utils/posts";
 import { type User, type Post } from "@prisma/client";
 import Link from "next/link";
 import PostFooter from "./PostFooter";
+import PostPreHeader from "./PostPreHeader";
 
 interface PostPreviewProps {
   post: Post;
@@ -10,17 +10,13 @@ interface PostPreviewProps {
 }
 
 export default function PostPreview({ post, user }: PostPreviewProps) {
-  const formattedDate = formatDate(post.updatedAt);
-
   return (
     <div className="container">
       <article
         key={post.id}
         className="[&:not(:last-child)]:border-b-1 border-gray-200"
       >
-        <time dateTime={formattedDate} className="text-gray-500 italic text-xs">
-          {formattedDate}
-        </time>
+        <PostPreHeader post={post} />
         <div className="group relative">
           <header>
             <h3 className="mt-3 text-3xl/8 font-semibold text-gray-900">
